# [performance](load) do not copy input_block in memtable