## Proposed changes

Pass `column_offset` to `Block::clone_without_columns()` and `MutableBlock::add_rows()`,
so we do not need to copy `input_block` in `MemTable::insert()`.

To optimize performance in cases with many (e.g. 12k+) tablets:

Before:

```
-  DeltaWriterWriteTime:  1m58s
   -  MemTableWriterFlushTime:  0ns
   -  MemTableWriterLockTime:  1s41ms
   -  MemTableWriterShrinkTime:  0ns
   -  MemTableWriterWriteTime:  1m53s
      -  MemTableCopyCount:  12.044003M  (12044003)
      -  MemTableCopyTime:  42s857ms
      -  MemTableCopyTime0:  11s866ms
      -  MemTableInitTime:  13.384ms
      -  MemTableInsertTime:  51s543ms
```

After:

```
-  DeltaWriterWriteTime:  1m
   -  MemTableWriterFlushTime:  0ns
   -  MemTableWriterLockTime:  971.481ms
   -  MemTableWriterShrinkTime:  0ns
   -  MemTableWriterWriteTime:  55s274ms
      -  MemTableCopyCount:  0
      -  MemTableCopyTime:  0ns
      -  MemTableCopyTime0:  0ns
      -  MemTableInitTime:  18.746ms
      -  MemTableInsertTime:  53s772ms
```

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...