@kaijchen 2024-06-28
run buildall